partial refactoring PyCourseSettingPanel
